Move pad items array to a seperate file
