+++ src/Data/padItems.js
@@ -1,0 +1,14 @@
+const padItems = [
+  { sym: 'AC', class: 'all_clear' },
+  { sym: '+/-', class: 'plus_neg' }, { sym: '%', class: 'modular' },
+  { sym: '/', class: 'operation divide' }, { sym: '7', class: 'seven' },
+  { sym: '8', class: 'eight' }, { sym: '9', class: 'nine' },
+  { sym: 'x', class: 'operation multiply' }, { sym: '4', class: 'four' },
+  { sym: '5', class: 'five' }, { sym: '6', class: 'six' },
+  { sym: '-', class: 'operation substract' }, { sym: '1', class: 'one' },
+  { sym: '2', class: 'two' }, { sym: '3', class: 'three' },
+  { sym: '+', class: 'operation add' }, { sym: '0', class: 'zero' },
+  { sym: '.', class: 'point' }, { sym: '=', class: 'operation equal' },
+];
+
+export default padItems;
